LayoutTests/fast/forms/input-null-expected.txt

This test setting various attributes of an input element to JavaScript null.

TEST SUCCEEDED: The value was the empty string. [tested input.defaultValue]
TEST SUCCEEDED: The value was the empty string. [tested input.accept]
TEST SUCCEEDED: The value was the empty string. [tested input.accessKey]
TEST SUCCEEDED: The value was the empty string. [tested input.align]
TEST SUCCEEDED: The value was the empty string. [tested input.alt]
TEST SUCCEEDED: The value was the empty string. [tested input.name]
TEST SUCCEEDED: The value was the empty string. [tested input.useMap]
TEST SUCCEEDED: The value was the empty string. [tested input.value]
TEST SUCCEEDED: The value was the empty string. [tested input.src]
